refactored tg clients into single instance
