+++ lib/teselagen/api/__init__.py
@@ -1,3 +1,4 @@
+from .client import TeselaGenClient
 from .discover_client import DISCOVERClient
 from .design_client import DESIGNClient
 from .test_client import TESTClient
